refactor: split some common functions
